refactor(bat): switch to scripts.call dispatcher and tighten script pattern input
add scripts.call.bat as the primary script dispatcher
update cezdev.bat startup calls to use scripts.call for libraries, environments, and projects
keep cezdev.crud.bat as a compatibility wrapper that forwards into scripts.call
require explicit scriptpattern in scripts.call and improve validation/error messages
simplify pattern resolution by using provided scriptpattern directly
preserve existing execution flow/logging while reducing legacy CRUD-specific coupling
