+++ ccore/c3dclasses/ccore/cconstructs/cvar/cmemory-test1.json
@@ -1,1 +1,1 @@
-{"var4":{"m_strname":"var4","m_iretrieved":1776544035508,"m_icreated":1776544035497,"m_strtype":"number","m_iupdated":-1,"m_value":6},"var2":{"m_strname":"var2","m_iretrieved":-1,"m_icreated":1776544035496,"m_strtype":"carray","m_iupdated":-1,"m_value":[1,2,3,4,5,6,7,8,9,10]},"var1":{"m_strname":"var1","m_iretrieved":-1,"m_icreated":1776544035496,"m_strtype":"string","m_iupdated":-1,"m_value":"var1 value"}}
+{"var4":{"m_strname":"var4","m_iretrieved":1776729356553,"m_icreated":1776729356543,"m_strtype":"number","m_iupdated":-1,"m_value":6},"var2":{"m_strname":"var2","m_iretrieved":-1,"m_icreated":1776729356543,"m_strtype":"carray","m_iupdated":-1,"m_value":[1,2,3,4,5,6,7,8,9,10]},"var1":{"m_strname":"var1","m_iretrieved":-1,"m_icreated":1776729356543,"m_strtype":"string","m_iupdated":-1,"m_value":"var1 value"}}
