+++ ccore/c3dclasses/csystem/cdevice/cmemory/cmemory-test.json
@@ -1,1 +1,1 @@
-{"kevin":{"m_strname":"kevin","m_iretrieved":1776544034108,"m_icreated":1776544034097,"m_strtype":"string","m_iupdated":-1,"m_value":"is here"},"charlie":{"m_strname":"charlie","m_iretrieved":1776544034139,"m_icreated":1776544034128,"m_strtype":null,"m_iupdated":-1,"m_value":"charlie and the chocolate man"}}
+{"kevin":{"m_strname":"kevin","m_iretrieved":1776729355402,"m_icreated":1776729355387,"m_strtype":"string","m_iupdated":-1,"m_value":"is here"},"charlie":{"m_strname":"charlie","m_iretrieved":1776729355434,"m_icreated":1776729355426,"m_strtype":null,"m_iupdated":-1,"m_value":"charlie and the chocolate man"}}
