+++ ccore/cezdev/cezdev.modified.bat
@@ -1,0 +1,8 @@
+@echo off
+echo [CALLING] %~nx0
+echo [PARAM] modtype: %~1
+echo [PARAM] modfile: %~2
+echo [PARAM] platform: %~3
+echo [PARAM] platform_name: %~4
+echo [ENDING] %~nx0
+
