+++ ccore/cezdev/junk/cenvironments.modified.bat
@@ -13,20 +13,21 @@ echo [CALLING] %~nx0
 
 set "MODTYPE=%~1"
 set "MODFILE=%~2"
-set "ITEMKIND=%~3"
-set "PLATFORM=%~4"
-set "PLATFORMNAME=%~5"
+set "PLATFORM=%~3"
+set "PLATFORMNAME=%~4"
 
 :: Extract filename from path
 for %%F in ("%MODFILE%") do set "FILENAME=%%~nxF"
 
 :: Log the modification
-echo [CENVIRONMENTS] %MODTYPE% %ITEMKIND%: %FILENAME% in %PLATFORMNAME%
+echo [CENVIRONMENTS] %MODTYPE%: %FILENAME% in %PLATFORMNAME%
 
 :: Look for environment-specific modified handler
 set "HANDLER=%CENVIRONMENTS%\%PLATFORMNAME%\%PLATFORMNAME%.modified.bat"
 if exist "!HANDLER!" (
-    call "!HANDLER!" "%MODTYPE%" "%MODFILE%" "%ITEMKIND%"
+    call "!HANDLER!" "%MODTYPE%" "%MODFILE%" "%PLATFORM%" "%PLATFORMNAME%"
+) else (
+    echo [INFO] Handler does not exist: %PLATFORMNAME%.modified.bat
 )
 
 :: Add custom handling logic here as needed
