+++ ccore/cezdev/junk/clibraries.crud.bat
