+++ ccore/cezdev/junk/clibraries.modified.bat
