+++ ccore/cezdev/junk/cprojects.modified.bat
