+++ cplatform/cenvironments/cjava/pom - Copy.xml
@@ -1,0 +1,158 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
+	http://maven.apache.org/maven-v4_0_0.xsd">
+	<properties>
+   <maven.compiler.source>1.8</maven.compiler.source>
+   <maven.compiler.target>1.8</maven.compiler.target>
+   </properties>
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>com.cezdev</groupId>
+  <artifactId>${release.artifactId}</artifactId>
+  <packaging>jar</packaging>
+  <version>${release.version}</version>
+  <name>cezdev</name>
+  <url>http://maven.apache.org</url>
+  
+  <dependencies>
+	<dependency>
+    	<groupId>com.googlecode.json-simple</groupId>
+    	<artifactId>json-simple</artifactId>
+    	<version>1.1.1</version>
+	</dependency>
+	<dependency>
+    	<groupId>commons-lang</groupId>
+    	<artifactId>commons-lang</artifactId>
+    	<version>2.6</version>
+	</dependency>
+	<dependency>
+  <groupId>junit</groupId>
+  <artifactId>junit</artifactId>
+  <version>4.12</version>
+</dependency>
+	<dependency>
+        <groupId>ProjectWeka</groupId>
+        <artifactId>weka</artifactId>
+        <scope>system</scope>
+        <version>1</version>
+    <systemPath>${release.path}\cplatform\clibraries\other\weka-3-7-12\weka.jar</systemPath>
+    </dependency>
+	<dependency>
+        <groupId>Jama</groupId>
+        <artifactId>jama</artifactId>
+        <scope>system</scope>
+        <version>1</version>
+    <systemPath>${release.path}\cplatform\clibraries\other\Jama-1.0.3\Jama-1.0.3.jar</systemPath>
+    </dependency>
+	<!-- https://mvnrepository.com/artifact/org.mockito/mockito-all -->
+	<!-- https://mvnrepository.com/artifact/org.mockito/mockito-core -->
+	<dependency>
+		<groupId>org.mockito</groupId>
+		<artifactId>mockito-core</artifactId>
+		<version>2.23.4</version>
+		<scope>test</scope>
+	</dependency>
+<!--
+	<dependency>
+    <groupId>com.amazonaws</groupId>
+    <artifactId>aws-java-sdk</artifactId>
+    <version>1.11.327</version>
+  </dependency>
+-->
+
+<dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-controls</artifactId>
+            <version>11</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-media</artifactId>
+            <version>11</version>
+            <classifier>win</classifier>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-media</artifactId>
+            <version>11</version>
+            <classifier>mac</classifier>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-media</artifactId>
+            <version>11</version>
+            <classifier>linux</classifier>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-graphics</artifactId>
+            <version>11</version>
+            <classifier>win</classifier>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-graphics</artifactId>
+            <version>11</version>
+            <classifier>mac</classifier>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-graphics</artifactId>
+            <version>11</version>
+            <classifier>linux</classifier>
+        </dependency>
+
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-fxml</artifactId>
+            <version>11</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.seleniumhq.selenium</groupId>
+            <artifactId>selenium-java</artifactId>
+            <version>4.0.0</version>
+        </dependency>
+
+  </dependencies>
+  <build>
+  <plugins>
+  <!-- Maven Assembly Plugin -->
+	<plugin>
+		<groupId>org.apache.maven.plugins</groupId>
+		<artifactId>maven-assembly-plugin</artifactId>
+		<version>2.4.1</version>
+		<configuration>
+			<!-- get all project dependencies -->
+			<descriptorRefs>
+				<descriptorRef>jar-with-dependencies</descriptorRef>
+			</descriptorRefs>
+		</configuration>
+		<executions>
+		  <execution>
+			<id>make-assembly</id>
+			 <!-- bind to the packaging phase -->
+			<phase>package</phase>
+			<goals>
+				<goal>single</goal>
+			</goals>
+		  </execution>
+		</executions>
+	</plugin>
+	<plugin>
+        <groupId>org.openjfx</groupId>
+        <artifactId>javafx-maven-plugin</artifactId>
+        <version>0.0.3</version>
+        <configuration>
+            <mainClass>HelloFX</mainClass>
+        </configuration>
+    </plugin>
+	</plugins>
+	</build>		
+</project>
